# Ebook Lint: Swap to Copying from Zip to Zip and Cut Out Temp File