--- a/ebook-lint/cmd/cbz/compress.go
+++ b/ebook-lint/cmd/cbz/compress.go
@@ -6,9 +6,7 @@ import (
 
 	"github.com/MakeNowJust/heredoc"
 	filesize "github.com/pjkaufman/go-go-gadgets/ebook-lint/internal/file-size"
-	"github.com/pjkaufman/go-go-gadgets/ebook-lint/internal/images"
 	filehandler "github.com/pjkaufman/go-go-gadgets/pkg/file-handler"
-	"github.com/pjkaufman/go-go-gadgets/pkg/image"
 	"github.com/pjkaufman/go-go-gadgets/pkg/logger"
 	"github.com/spf13/cobra"
 )
@@ -83,32 +81,20 @@ func init() {
 }
 
 func compressCbz(lintDir, cbz string) {
-	var src = filehandler.JoinPath(lintDir, cbz)
-	var dest = filehandler.JoinPath(lintDir, "cbz")
+	// var src = filehandler.JoinPath(lintDir, cbz)
+	// var dest = filehandler.JoinPath(lintDir, "cbz")
 
-	err := filehandler.UnzipRunOperationAndRezip(src, dest, func() error {
-		var imageFiles, err = filehandler.MustGetAllFilesWithExtsInASpecificFolderAndSubFolders(dest, image.CompressableImageExts...)
-		if err != nil {
-			return err
-		}
+	// filehandler.UnzipRunOperationAndRezip(src, dest, func() {
+	// 	var imageFiles = filehandler.MustGetAllFilesWithExtsInASpecificFolderAndSubFolders(dest, image.CompressableImageExts...)
 
-		for i, imageFile := range imageFiles {
-			if verbose {
-				logger.WriteInfof("%d of %d: compressing %q\n", i, len(imageFiles), imageFile)
-			}
+	// 	for i, imageFile := range imageFiles {
+	// 		if verbose {
+	// 			logger.WriteInfo(fmt.Sprintf(`%d of %d: compressing %q`, i, len(imageFiles), imageFile))
+	// 		}
 
-			err = images.CompressImage(imageFile)
-			if err != nil {
-				return err
-			}
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		logger.WriteError(err.Error())
-	}
+	// 		images.CompressImage(imageFile)
+	// 	}
+	// })
 }
 
 func ValidateCompressFlags(dir string) error {