--- a/ebook-lint/internal/epub-handler/parse-opf-contents.go
+++ b/ebook-lint/internal/epub-handler/parse-opf-contents.go
@@ -1,4 +1,4 @@
-package linter
+package epubhandler
 
 import (
 	"encoding/xml"