… ebook-lint/cmd/epub/common.go
… ebook-lint/cmd/epub/compress-and-lint.go
… ebook-lint/cmd/epub/lint-epub_test.go
… ebook-lint/cmd/epub/replace-strings.go
… ebook-lint/internal/epub-handler/parse-opf-contents_test.go
… ebook-lint/internal/epub-handler/update-epub.go
… ebook-lint/internal/images/compress-image.go
… ebook-lint/internal/images/compress-images.go
… ebook-lint/internal/zip-handler/update-zip.go
… git-helper/cmd/submodule-create.go
… git-helper/cmd/submodule-update.go
… jp-proc/cmd/proc.go
… magnum/internal/config/config-handler.go
… pkg/cmd-handler/add-generate-cmd.go
… pkg/file-handler/file-handler.go
… pkg/file-handler/zip.go
… song-converter/cmd/create-csv.go
… song-converter/cmd/create-html.go