Fixes #9 

When creating epub files, it makes sense to go ahead and not create a temporary file in order to make modifications and create a new epub.

Here is what the performance gains look like for a test I ran after all changes were said and done:
```
goos: linux
goarch: amd64
pkg: github.com/pjkaufman/go-go-gadgets/ebook-lint/cmd/epub
cpu: AMD Ryzen 5 3500U with Radeon Vega Mobile Gfx
           │ old2.0.txt  │          new2.0.txt          │
           │   sec/op    │   sec/op     vs base         │
LintEpub-8   179.7m ± 1%   160.1m ± 2%  -10.89% (n=100)
```

Todos:
- [x] Add functionality for CBZ image compression
- [x] Remove unused code
- [x] Retest and make sure performance improvements still stand